Refactor: App, i18n directories
- ConversationDetailPage -> RoomPage
- ConversationStartPage -> StartPage
- i18n 디렉토리 내부로 Locale 파일 위치 변경
- settings.json에 i18n 관련 경로 설정
